Extract the common data lake table url and options parsing for delta and iceberg scan functions
